requirements.txt
python-dotenv~=0.14.0
colorama~=0.4.4
aiohttp~=3.6.3
discord~=1.0.1
bs4~=0.0.1
beautifulsoup4~=4.9.3